Fix RuggedPOD API installation
Running a background command in the chroot context
fails with the current setup. It was actually not
necessary. This was only to be sure to write
periodically in the log file to avoid travis think
the build is stuck.

Replacing it by a dummy script that writes in the
log in a loop makes the trick even if it's pretty
dirty.